requirements.txt
pymupdf
llama-parse
python-dotenv
pytesseract
pdf2image
pdfplumber
